# doc/api.yaml
openapi: 3.0.3
info:
  description: |-
    Each user will be presented with a stream of photos in reverse chronological order, with information
    about when it was uploaded (date and time) and how many likes and comments it has. Users can place
    (and later remove) a “like” to photos from other users. Also, users can add comments to any photo
    (even those uploaded by themself). Comments can be removed by the author only.

    Users can ban other users. If Alice (user) bans Eve (user), Eve won’t be able to see any information
    about Alice. Alice can decide to remove the ban at any moment.

    Users will have their own personal profiles. The personal profile page for user detail: user’s photos (in
    reverse chronological order), their followers and following, and how many photos have been uploaded.
    Each user can change his/her own username, upload photos, remove photos, and follow/unfollow
    other users. Removal of a photo will also remove likes and comments.

    Users profiles can be searched via username.

    The login is perfomed by specifying the username.
  title: WASAPhoto API
  version: 1.0.0
  contact:
    name: Lucian Dorin Crainic
    email: crainic.1938430@studenti.uniroma1.it
    url: https://github.com/LucianCrainic
  license:
    name: MIT
...
paths:
...
  /user/{username}/stream:
    parameters:
      - $ref: "#/components/parameters/username"
    #* getMyStream
    get:
      security:
      - bearerAuth : []
      tags: ["User"]
      summary: Return user stream
      description: |-
        Return a list of photos for the user stream.
        If the client specifies a range for when the photos were uploaded,
        the list will be sorted by date from that point and filtered.
        Optionally, the client can specify a custom quantity of photos to be displayed.
      operationId: getMyStream
      parameters:
        - name: range
          in: query
          description: |-
            **Range** for when the photos were uploaded.
            Default: Today
          schema:
            description: When the photos were uploaded schema.
            type: string
            enum: ["Today","Yesterday", "Last week"]
            example: "Today"
        - name: quantity
          in: query
          required: false
          description: |-
            **Quantity** of photos to be displayed. 
            Default: 10.
          schema:
            description: Quantity schema.
            type: number
            format: float
            example: 15
            minimum: 1
            maximum: 50
      responses:
        "200": 
          description: User stream retrived successfully.
          content:
            image/*:
              schema: 
                type: array
                minItems: 0
                description: image array schema.
                maxItems: 999
                items: { $ref: "#/components/schemas/photosList"}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  #? get a user profile
...
  /users/{username}/photo:
    parameters:
      - $ref : "#/components/parameters/username"
    #* getUserPhotos
    get:
      security:
      - bearerAuth : []   
      tags: ["Photo"]
      summary: Return user photos
      operationId: getUserPhotos
      description: |-
        Return user photos, if the username provided exists. If the username does not exist then an error response is given.
      responses:
        "200": 
          description: User photos retrived.
          content:
            image/*:
              schema: 
                $ref: "#/components/schemas/photosList"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  #? like
...
  /users/{username}/photo/{photoId}/comment/{commentId}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/photoID"
      - $ref: "#/components/parameters/commentID"
    #* commentPhoto
    put:
      security:
      - bearerAuth : []
      tags: ["Comment"]
      summary: Add comment
      operationId: commentPhoto
      description: |-
       Add a comment under a Photo, if photoId is correct
       then the comment in the request body is added under the photo. If 
       the photoId is wrong then a response error is given.
      requestBody:
        description: Comment 
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/comment"
        required: true
      responses:
        "201":
          description:  Comment added successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/commentsList"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    #* uncommentPhoto
    delete:
     security:
     - bearerAuth : []
     tags: ["Comment"]
     operationId: uncommentPhoto
     summary: Remove comment
     description: |-
       Delete a comment under a photo. If the commentId is correct then the comment is removed from the photo, an error response
       is given if the commentId is not correct.
     responses:
       "200": 
        description: Comment removed successfully.
        content:
            application/json:
              schema: 
                $ref: "#/components/schemas/commentsList"
       "400": { $ref: "#/components/responses/BadRequest" }
       "401": { $ref: "#/components/responses/UnauthorizedError" }
       "500": { $ref: "#/components/responses/InternalServerError" }
  #?  get all coments under a photo
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
...
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
            schema:
              type: string
              description: Unique user identifier.
              pattern: '^.*?$'
              minLength: 1
              maxLength: 40
              example: Access token  missing.
    #* BadRequest response
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:  
        application/json:
            schema:
              type: string
              description: Unique user identifier.
              pattern: '^.*?$'
              minLength: 1
              maxLength: 40
              example: The request was not compliant.
    #* InternalServerError response
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
         application/json:
            schema:
              type: string
              description: Unique user identifier.
              pattern: '^.*?$'
              minLength: 1
              maxLength: 40
              example: internalError.
  schemas:
...
    commentsList: 
      title: Comments
      description: This object represent the comments under a single photo.
      type: object
      properties:
        photoId: 
          description:  Unique photo identifier.
          type: string
          pattern: '^.*?$'
          example: "#P1234"
          minLength: 3
          maxLength: 16
        comments: 
          description: list of objects (each object is a comment).
          type: array
          minItems: 0
          maxItems: 99999
          items:
            type: object
            properties:
              commentId:
                description: Unique comment identifier.
                type: string
                pattern: '^.*?$'
                example: "#C1234"
                minLength: 3
                maxLength: 16
              identifier:
                description: Unique user identifier.
                type: string
                pattern: '^.*?$'
                example: "#U5678" 
                minLength: 3
                maxLength: 16
              content:
                description: Content of the comment.
                type: string
                pattern: '^.*?$'
                example: "Hello World!"
                minLength: 2
                maxLength: 30
    #* likesList schema
...
    photosList: 
      title: Photos
      description: This object represent the photos of a single user.
      type: object
      properties:
        identifier:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234" 
          minLength: 3
          maxLength: 16
        photos: 
          description: list of objects (each object is a photo)
          type: array
          minItems: 0
          maxItems: 99999
          items:
            type: object
            properties:
              photoId: 
                description:  Unique photo identifier.
                type: string
                pattern: '^.*?$'
                example: "#P1234"
                minLength: 3
                maxLength: 16
              file: 
                description: file
                type: string
                format: binary
                minLength: 3
                maxLength: 16
    #* login schema
...
    comment:
      type: object
      description: content schema.
      properties:
        content:
          description: content
          type: string
          example: Hello World!
          pattern: '^.*?$'
          minLength: 1
          maxLength: 40
  parameters:
...
    username:
      schema:
          description: username schema.
          type: string
          pattern: '^.*?$'
          example: LucianCrainic
          minLength: 5
          maxLength: 30
      name: username
      in: path
      required: true
      description: "**Username** of a WASAPhoto user."
    #* photoID parameter
    photoID:
      schema:
        description: photoID schema.
        type: string
        pattern: '^.*?$'
        example: "#P1234"
        minLength: 3
        maxLength: 16
      name: photoId 
      in: path
      required: true
      description: "**photoID** of a specific photo on a user profile."
    #* commentID parameter
    commentID:
      schema:
        description: commentID schema.
        type: string
        pattern: '^.*?$'
        example: "#C1234"
        minLength: 3
        maxLength: 16
      name: commentId
      in: path
      required: true
      description: "**commentID** of a single comment under a photo."
    #* likeID parameter
...